// types/user-ekyc/education.ts
export type EducationInfoType = {
  certifications?: string
  yearAttended: string,
  highestEducationLevel?: string
  institution?: string
}
export type CreateEducationInfoType = {
  certifications?: string
  yearAttended: string,
  highestEducationLevel?: string
  institution?: string
}

export type UpdateEducationInfoType = {
  userId?: string
  certifications?: string
  yearAttended?: string,
  highestEducationLevel?: string
  institution?: string
}